Moved SectionTitle to portfolio
