+++ src/components/portfolio/Skills.tsx
@@ -1,6 +1,6 @@
 import skills from "@/data/skills";
 import Image from "next/image";
-import SectionTitle from "@/components/ui/SectionTitle";
+import SectionTitle from "@/components/portfolio/SectionTitle";
 
 export default function Skills() {
   const Item = ({imageUrl, name, description}: {imageUrl: string, name: string, description: string}) => {
@@ -16,7 +16,7 @@ export default function Skills() {
   }
 
   return (
-    <section id="about">
+    <section id="about" className="mb-9">
       <SectionTitle title="Skills" />
       <p className="my-2">I&#39;ve worked with a variety of languages and frameworks. These are some of them I focus on.<br/> You can view the rest of my skills on my resume :)</p>
       <div className="grid md:grid-cols-4 grid-cols-2 gap-4">
